+++ R/eQTLcatalogue_fetch.R
@@ -2,10 +2,10 @@
 #'
 #' Query eQTL Catalogue datasets with multiple methods options.
 #' 
-#' @param unique_id Unique eQTL Catalogue ID assigned in metadata
-#'  ("unique_id" column in \code{data(meta)}). 
+#' @param dataset_id Unique eQTL Catalogue ID assigned in metadata
+#'  ("dataset_id" column in \code{data(meta)}). 
 #' @param query_dat \link[data.table]{data.table} of GWAS summary statistics.
-#' @param add_qtl_id Add "qtl_id" (i.e. "unique_id") column to the query result.
+#' @param add_dataset_id Add "dataset_id" (i.e. "dataset_id") column to the query result.
 #' @param convert_genes Convert Ensembl IDs to HGNC symbols.
 #' @param chrom Chromosome of the query window.
 #' @param bp_lower Minimum basepair position of the query window.
@@ -20,30 +20,26 @@
 #' query_granges <- echodata::BST1
 #' GWAS.QTL_manual <- catalogueR::eQTLcatalogue_fetch(
 #'   query_granges = query_granges,
-#'   unique_id = meta$unique_id[1])
-eQTLcatalogue_fetch <- function(unique_id,
+#'   dataset_id = meta$dataset_id[1])
+eQTLcatalogue_fetch <- function(dataset_id,
                                 query_granges,
                                 method = c("REST","tabix"),
                                 quant_method = "ge", 
                                 multithread_tabix = FALSE,
-                                add_qtl_id = TRUE,
+                                add_dataset_id = TRUE,
                                 convert_genes = TRUE,
                                 suffix = ".QTL",
                                 timeout = 5*60,
                                 conda_env = "echoR_mini",
                                 nThread = 1, 
-                                verbose = TRUE) {
-  
-  # echoverseTemplate:::args2vars(catalogueR:::eQTLcatalogue_fetch)
-  # echoverseTemplate:::source_all()
-  
-  qtl_id <- NULL;
+                                verbose = TRUE) { 
+  dataset_id <- NULL;
   method <- tolower(method)[1]  
   if (method=="tabix") {
       ## Tabix is about ~17x faster than the REST API,
       ## but gets blocked often by EBI server which thinks its an attack.
     qtl.subset <- fetch_tabix(
-          unique_id = unique_id,
+          dataset_id = dataset_id,
           query_granges = query_granges, 
           quant_method = quant_method,  
           nThread = if (isTRUE(multithread_tabix)) nThread else 1,
@@ -51,7 +47,7 @@ eQTLcatalogue_fetch <- function(unique_id,
           verbose = verbose)
   } else {
     qtl.subset <- fetch_restAPI(
-          unique_id = unique_id,
+          dataset_id = dataset_id,
           quant_method = quant_method, 
           query_granges = query_granges, 
           timeout = timeout,
@@ -62,8 +58,8 @@ eQTLcatalogue_fetch <- function(unique_id,
     colnames(qtl.subset) <- paste0(colnames(qtl.subset), suffix)
   } 
   #### Post=processing ####
-  if (isTRUE(add_qtl_id)) {
-    qtl.subset[,qtl_id:=unique_id]
+  if (isTRUE(add_dataset_id)) {
+    qtl.subset[,dataset_id:=dataset_id]
   }
   #### Convert genes ####
   if (isTRUE(convert_genes)) {