+++ R/eQTLcatalogue_iterate_fetch.R
@@ -5,10 +5,10 @@
 #' @source
 #' \code{
 #' sumstats_paths <- echodata::get_Nalls2019_loci(limit_snps = 5)
-#' qtl_id <- catalogueR::eQTLcatalogue_list_datasets()$unique_id[1]
+#' dataset_id <- catalogueR::eQTLcatalogue_list_datasets()$unique_label[1]
 #' GWAS.QTL <- catalogueR:::eQTLcatalogue_iterate_fetch(
 #'     sumstats_paths = sumstats_paths,
-#'      qtl_id = qtl_id,
+#'      dataset_id = dataset_id,
 #'      nThread = 1,
 #'      split_files = FALSE)
 #' }
@@ -23,7 +23,7 @@ eQTLcatalogue_iterate_fetch <- function(sumstats_paths,
                                             tempdir(),
                                             "catalogueR_queries"
                                         ),
-                                        qtl_id,
+                                        dataset_id,
                                         method = c("REST","tabix"),
                                         quant_method = "ge",
                                         multithread_loci = TRUE,
@@ -79,7 +79,7 @@ eQTLcatalogue_iterate_fetch <- function(sumstats_paths,
         # Test if query file already exists
         split_path <- make_split_path(
             output_dir = output_dir,
-            qtl_id = qtl_id,
+            dataset_id = dataset_id,
             loc = loc
         )
         if (file.exists(split_path) && force_new_subset == FALSE) {
@@ -104,12 +104,12 @@ eQTLcatalogue_iterate_fetch <- function(sumstats_paths,
             qtl.subset <- tryCatch(
                 expr = {
                     eQTLcatalogue_fetch(
-                        unique_id = qtl_id,
+                        dataset_id = dataset_id,
                         query_granges = query_dat,
                         method = method,
                         quant_method = quant_method,
                         multithread_tabix = multithread_tabix, 
-                        add_qtl_id = TRUE,
+                        add_dataset_id = TRUE,
                         convert_genes = TRUE,
                         nThread = nThread,
                         conda_env = conda_env,