+++ tests/testthat/test-eQTLcatalogue_fetch.R
@@ -1,13 +1,13 @@
 test_that("eQTLcatalogue_fetch works", {
     
     data("meta")
-    unique_id <- meta$unique_id[1]
+    unique_label <- meta$unique_label[1]
     paths <- echodata::get_Nalls2019_loci(limit_snps = 5)
     query_dat <- data.table::fread(paths$BST1)
     
     #### Using data ####
     gwas.qtl <- catalogueR::eQTLcatalogue_fetch(
-        unique_id = unique_id,
+        unique_label = unique_label,
         query_granges = query_dat, 
         method = "REST"
     )
@@ -20,7 +20,7 @@ test_that("eQTLcatalogue_fetch works", {
       query_start_pos = min(query_dat$POS), 
       query_end_pos = max(query_dat$POS))
     GWAS.QTL_manual <- catalogueR::eQTLcatalogue_fetch(
-      unique_id = unique_id, 
+      unique_label = unique_label, 
       query_granges = query_granges)
     testthat::expect_true(methods::is(GWAS.QTL_manual, "data.table"))
     testthat::expect_gte(nrow(GWAS.QTL_manual), 1200)
@@ -34,7 +34,7 @@ test_that("eQTLcatalogue_fetch works", {
     ## WARNING: if you test this iteratively, will eventually get blocked 
     ## by the EMBL-EBI server.
     gwas.qtl <- catalogueR::eQTLcatalogue_fetch(
-      unique_id = unique_id,
+      unique_label = unique_label,
       query_granges = query_dat, 
       method = "tabix"
     )